Move torch/tf checkpoint tests back to predictor test file
