+++ python/ray/train/tests/test_torch_predictor.py
@@ -54,6 +54,16 @@ class DummyCustomModel(torch.nn.Module):
         return [input_tensor, input_tensor]
 
 
+def assert_equal_torch_models(model1, model2):
+    # Check equality by comparing their `state_dict`
+    model1_state = model1.state_dict()
+    model2_state = model2.state_dict()
+    assert len(model1_state.keys()) == len(model2_state.keys())
+    for key in model1_state:
+        assert key in model2_state
+        assert torch.equal(model1_state[key], model2_state[key])
+
+
 @pytest.fixture
 def model():
     return DummyModelSingleTensor()
@@ -84,6 +94,18 @@ def test_init(model, preprocessor):
 
     assert checkpoint_predictor.model == predictor.model
     assert checkpoint_predictor.get_preprocessor() == predictor.get_preprocessor()
+
+
+def test_torch_checkpoint():
+    model = torch.nn.Linear(1, 1)
+    checkpoint = TorchCheckpoint.from_model(model)
+    assert_equal_torch_models(checkpoint.get_model(), model)
+
+    with checkpoint.as_directory() as path:
+        checkpoint = TorchCheckpoint.from_directory(path)
+        checkpoint_model = checkpoint.get_model()
+
+    assert_equal_torch_models(checkpoint_model, model)
 
 
 @pytest.mark.parametrize("use_gpu", [False, True])
